feat: move Tabs to separate folder for readability
