+++ packages/vscode-extension/src/network/components/Tabs/HeadersTab.css
