+++ packages/vscode-extension/src/network/components/Tabs/HeadersTab.tsx
@@ -1,6 +1,6 @@
 import "./HeadersTab.css";
 import { VscodeCollapsible } from "@vscode-elements/react-elements";
-import { NetworkLog } from "../hooks/useNetworkTracker";
+import { NetworkLog } from "../../hooks/useNetworkTracker";
 
 interface HeadersTabProps {
   networkLog: NetworkLog;
