+++ packages/vscode-extension/src/network/components/Tabs/PayloadAndResponseTab.css
